+++ app/src/main/java/com/example/coronadiagnosticapp/ui/fragments/questions/QuestionFragment.kt
@@ -12,6 +12,10 @@ import com.example.coronadiagnosticapp.R
 import com.example.coronadiagnosticapp.data.db.entity.Question
 import com.example.coronadiagnosticapp.data.db.entity.QuestionType.MULTI_SELECT
 import com.example.coronadiagnosticapp.data.db.entity.QuestionType.SELECT
+import com.example.coronadiagnosticapp.ui.fragments.questions.adapters.MultiQuestionAdapter
+import com.example.coronadiagnosticapp.ui.fragments.questions.adapters.SelectQuestionAdapter
+import com.example.coronadiagnosticapp.ui.fragments.questions.adapters.Selectable
+import com.example.coronadiagnosticapp.ui.fragments.questions.viewmodels.QuestionFragmentViewModel
 import com.example.coronadiagnosticapp.utils.getAppComponent
 import com.example.coronadiagnosticapp.utils.toast
 import kotlinx.android.synthetic.main.fragment_question.*
@@ -44,7 +48,10 @@ class QuestionFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        options_rv.adapter = MultiQuestionAdapter(emptyList())
+        options_rv.adapter =
+            MultiQuestionAdapter(
+                emptyList()
+            )
         doTheFlow()
     }
 
@@ -71,8 +78,12 @@ class QuestionFragment : Fragment() {
         val options = question.extraData
 
         val adapter: RecyclerView.Adapter<*> = when (question.type) {
-            MULTI_SELECT -> MultiQuestionAdapter(options)
-            SELECT -> SelectQuestionAdapter(options)
+            MULTI_SELECT -> MultiQuestionAdapter(
+                options
+            )
+            SELECT -> SelectQuestionAdapter(
+                options
+            )
             else -> throw IllegalArgumentException("Not the right type of question")
         }
