+++ app/src/main/java/com/example/coronadiagnosticapp/ui/fragments/questions/viewholders/BaseViewHolder.kt
@@ -1,4 +1,4 @@
-package com.example.coronadiagnosticapp.ui.fragments.questions
+package com.example.coronadiagnosticapp.ui.fragments.questions.viewholders
 
 import android.view.LayoutInflater
 import android.view.View
@@ -7,7 +7,12 @@ import androidx.annotation.LayoutRes
 import androidx.recyclerview.widget.RecyclerView
 
 abstract class BaseViewHolder(viewGroup: ViewGroup, @LayoutRes layout: Int) :
-    RecyclerView.ViewHolder(inflate(viewGroup, layout)) {
+    RecyclerView.ViewHolder(
+        inflate(
+            viewGroup,
+            layout
+        )
+    ) {
 
     companion object {
         private fun inflate(viewGroup: ViewGroup, layout: Int): View {
